Разбил TypeRepository на каждый отдельный репозиторий
