--- CookbookWebApplication/KitBook/Models/Repositories/Interfaces/ITypeRepository.cs
@@ -1,12 +1,0 @@
-﻿using KitBook.Models.Database.Entities.Types;
-
-namespace KitBook.Models.Repositories.Interfaces
-{
-    public interface ITypeRepository :
-        IRepository<RecipeType>,
-        IRepository<CookingType>,
-        IRepository<IngredientType>,
-        IRepository<DishType>
-    {
-    }
-}
